fix: address adversarial-review findings
- production_safety: latch the flatten-submitted flag only after every
  intent submits successfully; a failed submit previously latched and
  the breached position stayed on with no retry
- FillRecord.qty int -> float; int() truncated fractional (crypto)
  fills to zero in the ledger
- crypto recorders: open session files one at a time into the assigned
  dict; a mid-comprehension open() failure leaked the already-opened
  handles because the dict assignment never completed

+++ packages/backtest_pipeline/src/backtest_result.py
@@ -16,7 +16,9 @@ class FillRecord:
     timestamp_ns: int
     side: str
     exec_price: float
-    qty: int
+    # float, not int: crypto fills are fractional and int() would truncate
+    # a 0.5-contract fill to zero in the ledger.
+    qty: float
     hypothesis_id: int
     signal: float
     mid_at_exec: float
